Added the third model, with iso only on the body.
Changed the train, now it swich following the user's choice.
Created the new sub-class of Dataset, for the third model: dataset_body.
Now in processing the user choose which dataset and model wants to use.
Minor update in the cnns class, the dimension tensor.ones is output.
Created the cnn sub-class for the only iso on body model.
Adapted visualization_cnn to the body model.